feat: provide a release verification process
The now defunct website 911datasets.org made a series of 'releases' available with all sorts of data
related to the terrorist attacks of September 11, 2001. Each release came in the form of a torrent.
The original torrents now have either very little activity, or none at all. However, some of the
material is still available at various different places. The intention of this repository is to
provide tools to collect and verify the data.

This initial commit provides a verification process for a release. The idea is to use the SHA1
hashes in the torrent file to verify that the files obtained from other locations actually match the
original data. A torrent file consists of a list of file paths and a list of 'pieces' of a certain
size, where each piece is a SHA1 hash. The pieces relate to the content as a whole, rather than
individual files within the torrent.

We can verify that release files obtained from some other location match what was provided by the
torrents. You point the program at a directory of files that constitute a release, where the
directory has the same tree as the files in the release torrent. You then check they match what was
provided by the torrent. This is done by iterating over all the pieces in the torrent and comparing
them against hashes of the content of the files in the directory. For small files, the pieces can
span multiple files; for large files, thousands of pieces can relate to the same file. If you have
all the files and the SHA1 hashes of all their collective bytes match the piece hashes, you have
verified that the content you have obtained at some other location is the *exact* same as what was
provided by the torrent. Otherwise, you can indicate what files are missing or corrupted, and
attempt to retrieve them from somewhere else.

The process in this commit seems to work. It was developed with help from ChatGPT.

This commit is a bit of a mess, with hard coded values and such, but this will all get tidied up
in subsequent commits, when a proper interface is added.

There is also some code in here for building a list of releases based on the content of a release
table file and the torrent files, but this is going to be changed to a process which will store the
releases in a SQLite database.