feat: provide `init` command
The releases are parsed from static data embedded in the binary, then their torrent files are read,
which provides us with the file count and size information. There are a few releases with torrent
links that unfortunately are dead. The releases are then saved in the database. At this point, the
verification status of the releases are not known.