feat: provide a `verify` command
I've already described the verification process in the first commit message, so I won't repeat that
here. This commit provides a command to run the verification and save the result in the database.

In the future, something like a `--force` argument will have to be provided that allows a
verification to run again, e.g., for missing releases that have now been obtained.