feat: provide `--directory` flag for `ls` command
Print out the name of the directory rather than the name of the release, which can be helpful for
checking verification outcomes, since the name of the release and directory can be completely
different.

BREAKING CHANGE: this involved saving the directory name in the database, which meant I had to
delete my previous database. I could have added it as a 'migration', but I since I haven't verified
too many releases so far I decided I was fine with just deleting my current database and building a
new one.