feat: provide release download mechanism
Some releases on the archive have the same tree as the torrents. These can be downloaded using the
primitive mechanism provided by this command. It downloads each file sequentially and only shows
progress in terms of files downloaded vs the number of files in the torrent.

There are also a few bug fixes here with respect to reporting on the status of releases and resets
of verification outcomes.