chore: verification process bails early
My original idea for verification was to attempt to verify the whole release, and in the process
identify any files that were corrupted or missing. I've came to realise that doesn't really work.

First of all, if any file is missing, we can't verify the release, so there is no point trying. For
that reason, the process is changed to look for any missing files as a first step. If any files are
missing, we return with an incomplete result; if they all are, we return with a missing result--that
is to say, the whole release is missing.

Secondly, if a hash mismatch is discovered with a piece, we bail out without bothering to check the
rest of the release. We provide the file index at the point where the mismatch occurred. For one of
the releases, I have a mismatch, but the file at the given index doesn't seem to be corrupted. I'm
not sure what's going on there.

Also provide a command for dumping out the list of files in the release torrent.