feat: store torrents in the database
A database table is now used to store the torrent content. The torrents are downloaded as part of
the initialisation process.

All the `--torrents-path` arguments were removed.

I also uploaded all the torrents to one of my own S3 buckets and used that as the source of the
downloads, because sometimes the download links from the Wayback Machine are just painfully slow.