feat: store download url in the database
This is something that doesn't really need to vary and can just be provided in the static data.

There still needs to be a distinction made between releases that point to zip files and those that
have the directory tree, but at the moment only the latter was supported for downloading anyway.