feat: provide `mark-missing` command
The large 36411 FFA Records release had to be manually marked as missing because it has a small text
file in common with the uncompressed release, which technically causes the release to be marked as
incomplete. Since we have no other files, this release really should just be marked as missing.