feat: download uncompressed release 14
This release needs to be treated as a special case, because the data is scattered across different
collections in the Archive. The links have been provided in the form of static hash table, which
is then saved in the database.

For each path in the tree, it looks up its parent directory in the hash table to get the URL of the
collection where the file is. The file path is then appended on to the URL.