feat: check sizes even if files are missing
The `check` command was exiting if it detected any missing files. For release 14, we know that will
have some missing files, but we still want to check the file sizes of the files that do exist, so we
now give the command the opportunity to do that without exiting if it detects any missing files.