fix: deal with release 14 sub directories
A few of the directories within release 14 also have sub directories. This is a bit of a hacky
solution that also checks not just the parent directory of the current file, but the grand parent.
In the second situation, we're dealing with a sub directory situation.

Also fixes the invalid link to the New York Firefighters documentary.